+++ CHANGELOG.md
@@ -13,7 +13,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   [zedokai](https://github.com/slymax/zedokai) themes (#371) by @koru
 - Add notes about `ethical-foss/open-slopware` (#376) by @koru
 - Add option to hide the title bar (#375) (zed#37428) by @ycna07
-- node_runtime: Fix npm v12 output deserialization (#381) (zed#60798)
 
 ## Changed
 
@@ -27,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update root_path for compatibility with language servers (zed#48587) (#362)
 - Register toggle inline values action (zed#58921) (#373)
 - Fix emoji rendering on macOS (#378) from @Petrosz007
+- Fix npm v12 output deserialization (#381) (zed#60798)
 
 ## [3.0.1] - 2026-06-30